Use Flakes
This allows hydra to build via release.nix, with an override that
replaces nixpkgs with the hydra input.

This is a bit of a hack, and should probably be replaced by use of the
not-yet-sufficient Flakes support in Hydra once it can do things like
overriding and updating flake inputs.